Add http read/write response tests
