+++ server/Tests/adult-registration.request
@@ -4,7 +4,7 @@ Accept-Encoding: gzip, deflate, br
 Accept-Language: en,de-DE;q=0.9,de;q=0.8,en-US;q=0.7
 Cache-Control: no-cache
 Connection: keep-alive
-Content-Length: 643
+Content-Length: 618
 Content-Type: multipart/form-data; boundary=----WebKitFormBoundaryiB5iskbmcAfH1zPo
 DNT: 1
 Host: localhost:3000
